Add 'sf_noahmp' to list of possible_values for config_lsm_scheme
This commit adds 'sf_noahmp' to the list of possible options for the
config_lsm_scheme namelist variable in the atmosphere core's Registry.xml file.
Also included in this commit is an update of the name of the 'noah' option to
'sf_noah' under possible_values for config_lsm_scheme.