Restore CMake support for building the init_atmosphere core
This commit restores CMake support for the atmosphere core, which broke
due to changes made prior to the version 8.2 release. The breaking
changes were a result of the addition of mpas_init_atm_thompson_aerosols.F
to the init_atmosphere core without updating the CMakeLists.txt. To fix
this problem, this file was appended to the list of source files in the
init_atmosphere core's CMakeLists.txt.