Copy invariant fields used in atm_set_smlstep_pert_variables_work with OpenACC
Ensure that the invariant fields that are used in this work routine are
present on the device from model startup to model shutdown.

Fields are added to the device with copyin directives and other code
within mpas_atm_dynamics_init and fields are deleted from the device
with matching delete directives and code in mpas_atm_dynamics_finalize.