Make Noah-MP-only static fields optional in init_atmosphere core
Prior to this commit, the soilcomp and soilcl{1,2,3,4} fields needed only by the
Noah-MP scheme were required in all static files read by the init_atmosphere
core. In cases where the Noah-MP scheme will not be used when running the
atmosphere core, it would be desirable to not require these Noah-MP-only fields.

This commit introduces a new namelist option and package in the init_atmosphere
core to permit the use of static files that do not contain soilcomp and
soilcl{1,2,3,4}.

1) The new config_noahmp_static namelist option in the &data_sources namelist
group now determines whether the Noah-MP-only static fields will be allocated,
read, computed, and written in the init_atmosphere core. This namelist option
defaults to true, yielding the default behavior of the init_atmosphere_model
program prior to this commit.

2) The new noahmp package, which is active if and only if config_noahmp_static
is true, controls the allocation, reading, and writing of the Noah-MP-only
static fields.

3) In the init_atm_static routine, which is called when config_static_interp is
true, the soilcomp and soilcl{1,2,3,4} fields are computed only if
config_noahmp_static is true.