Fix memory leak from cellLimitField in mpas_block_creator_build_cell_halos
The cellLimitField field is allocated in the mpas_block_creator_build_cell_halos
routine and is only used in a call therein to mpas_dmpar_get_exch_list. If not
deallocated at the end of the mpas_block_creator_build_cell_halos routine, the
memory allocated to cellLimitField would be otherwise be leaked, since there
is nothing outisde of the mpas_block_creator_build_cell_halos routine that
retains a reference to it and deallocates it.

This commit fixes this memory leak by deallocating cellLimitField before the
mpas_block_creator_build_cell_halos routine returns.