Add unit tests to the test core for mpas_adjust_alarm_to_reference_time
This commit adds unit tests for the mpas_adjust_alarm_to_reference_time routine
to the test_core_timekeeping_tests module in the test core.

Implied in these tests is the requirement that, after a call to
mpas_adjust_alarm_to_reference_time, the prevRingTime of the alarm is set to a
value that ensures that an immediately subsequent call to mpas_is_alarm_ringing
will show that the alarm is ringing.

In order to get more verbose printout, including the internal state of alarms,
the MPAS_ADJUST_ALARM_VERBOSE macro can be modfied near the top of the
test_core_timekeeping_tests module so that its argument is not turned into a
comment.