Consolidating access to the config_dt namelist option via atm_setup_clock
This commit introduces changes to the atmosphere core in order to ensure that
the config_dt namelist option is accessed only once during the model run inside
atm_simulation_clock_init called from the atm_setup_clock function. Therafter,
the various parts of the atmosphere core are expected to retrieve the timestep
size from either the clock or via an argument list. This set of changes are in
preparation for allowing variable timestep sizes in the MPAS atmosphere core.