Eliminate data copies between mpas_atm_get_bdy_tend and atm_bdy_adjust_dynamics_speczone_tend
The ru_driving_tend, rt_driving_tend, and rho_driving_tend allocatable arrays
in the atm_time_integration module are now automatically allocated on the device
through '!$acc declare create' directives when MPAS-A is run with OpenACC
enabled. This allows the 'create' and 'copyout' of these variables to be
eliminated from the mpas_atm_get_bdy_tend routine, and it also allows the
'copyin' and 'delete' of these variables to be eliminated from the
atm_bdy_adjust_dynamics_speczone_tend routine.