Connect generated package logic capability to the atmosphere core
In order for the atmosphere core to make use of the ability for package
logic to be generated by the registry through the active_when XML attribute,
this commit includes the generated code in 'setup_packages.inc' in the
atm_core_interface module, and it also adds a call to the generated
atm_setup_packages_when routine from the atm_setup_packages routine.

Note that the call to the generated atm_setup_packages_when routine occurs
before any of the hand-written logic in the atm_setup_packages routine. As a
consequence, settings of packages in the generated logic will be superseded by
settings from hand-written logic for those packages.