Update MMM-physics tag in Externals.cfg to fix .F90 re-compilation issue
This commit updates the tag to 20250616-MPASv8.3 for the MMM-physics external in
the src/core_atmosphere/Externals.cfg file to address an issue with .F90 files
not being re-compiled to .o files if a .F90 file was modified. With the updated
tag, compiling the atmosphere core, then making changes to any of the .F90 files
in src/core_atmosphere/physics/physics_mmm/, then running 'make' again (without
first cleaning) leads to the modified .F90 files correctly being re-compiled.